r600g/compute: Avoid defragmenting the pool when it is not fragmented
This commit avoids defragmenting the pool when it isn't fragmented
at all. This can happens when you have more than two items in it
and free some items from the middle until the end.

For example, a pool with three items, A, B and C, and we free B and C.
When freeing B, the pool will be marked as fragmented, but when we
free C, the pool is now not fragmented, but still marked as being.

There's three ways of avoiding this:
1) Checking in compute_memory_defrag whether the pool is fragmented
2) Checking in compute_memory_finalize_pending whether the pool
    is fragmented when calculating the allocated size
3) Checking in compute_memory_finalize_pending whether the pool
    is fragmented after calculating the allocated size comparing
    the position a new item will have with the allocated size.

This patch implements the first one